refactor: Create validateIngredientsQuery file
