+++ recipe-app/src/components/SideBar.jsx
@@ -1,6 +1,7 @@
 // import data from "../data/recipes";
-import { AiOutlineSearch } from "react-icons/ai";
 import React, { useState } from "react";
+import validateIngredientsQuery from "../validateIngredientsQuery";
+import { AiOutlineSearch } from "react-icons/ai";
 
 export default function SideBar() {
   const [menu, SetMenu] = useState("nav-container");
@@ -10,71 +11,9 @@ export default function SideBar() {
   // });
   // console.log(recipes);
 
-  function handleSearchInputChange(e) {
-    setSearchedText(e.target.value);
-  }
-
-  function handleSearchBtnClick() {
-    if (/^[^_|\W]/.test(searchedText)) {
-      const errorModal = document.getElementById("error-modal");
-      const url =
-        "https://tasty.p.rapidapi.com/recipes/list?from=0&size=1&tags=under_30_minutes";
-      const options = {
-        method: "GET",
-        headers: {
-          "X-RapidAPI-Key": import.meta.env.VITE_X_RAPIDAPI_KEY,
-          "X-RapidAPI-Host": "tasty.p.rapidapi.com",
-        },
-      };
-
-      errorModal.style.display = "none";
-
-      async function checkIfRecipeExistsForIngredients() {
-        try {
-          let searchedWord = null;
-          const response = await fetch(url, options);
-          const result = await response.text();
-          const recipesArray = JSON.parse(result).results;
-          const hasSearchedIngredients = findIngredients();
-
-          function findIngredients() {
-            const searchedWordsArray = searchedText
-              .toLowerCase()
-              .trim()
-              .split(/[\W|_]/g)
-              .filter((i) => i);
-
-            console.log(searchedWordsArray);
-
-            return searchedWordsArray.some((w) => {
-              searchedWord = w;
-              return recipesArray.some((r) => {
-                return r.sections[0].components.some((c) => {
-                  console.log(c.ingredient.name);
-                  const regExpVersionOfSearchedWord = new RegExp(w);
-                  return regExpVersionOfSearchedWord.test(c.ingredient.name);
-                });
-              });
-            });
-          }
-
-          console.log(hasSearchedIngredients);
-          if (!hasSearchedIngredients) {
-            const ingredient404Element =
-              document.getElementById("ingredient-404");
-
-            ingredient404Element.innerText = searchedWord;
-            errorModal.style.display = "flex";
-            console.log(searchedWord);
-          }
-          console.log(recipesArray);
-        } catch (error) {
-          console.error(error);
-        }
-      }
-      console.log(searchedText);
-      checkIfRecipeExistsForIngredients();
-    }
+  function handleSearchBtnClick(searchedText) {
+    validateIngredientsQuery(searchedText);
+    setSearchedText("");
   }
 
   // function showSearch() {}
@@ -99,9 +38,12 @@ export default function SideBar() {
             placeholder="Enter Ingredient"
             className="search-box"
             value={searchedText}
-            onChange={handleSearchInputChange}
+            onChange={(e) => setSearchedText(e.target.value)}
           />
-          <button className="search-btn" onClick={handleSearchBtnClick}>
+          <button
+            className="search-btn"
+            onClick={() => handleSearchBtnClick(searchedText)}
+          >
             <AiOutlineSearch />
           </button>
         </div>
